refactor: use APM analyitics
